Change how splits are started
This patch changes the way in which splits are started upon creating a
new game. Instead of relying on a value dependent on the starting room
that also seemed somewhat inconsistent across multiple runs, the
autosplitter will now just look for the change in room ID from the
opening sequence to the starting room. This has the added advantage that
splits will no longer start upon loading an existing file.